将顶栏改为侧边栏
